Add separate base64_util module
