+++ lib/remap-functions/src/encode_base64.rs
@@ -1,3 +1,4 @@
+use crate::base64_util::Charset;
 use remap::prelude::*;
 use std::str::FromStr;
 
@@ -39,43 +40,6 @@ impl Function for EncodeBase64 {
             padding,
             charset,
         }))
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum Charset {
-    Standard,
-    UrlSafe,
-}
-
-impl Default for Charset {
-    fn default() -> Self {
-        Self::Standard
-    }
-}
-
-impl Into<base64::CharacterSet> for Charset {
-    fn into(self) -> base64::CharacterSet {
-        use Charset::*;
-
-        match self {
-            Standard => base64::CharacterSet::Standard,
-            UrlSafe => base64::CharacterSet::UrlSafe,
-        }
-    }
-}
-
-impl FromStr for Charset {
-    type Err = &'static str;
-
-    fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
-        use Charset::*;
-
-        match s {
-            "standard" => Ok(Standard),
-            "url_safe" => Ok(UrlSafe),
-            _ => Err("unknown charset"),
-        }
     }
 }
