+++ lib/remap-functions/src/lib.rs
@@ -1,4 +1,5 @@
 mod util;
+mod base64_util;
 
 #[cfg(feature = "append")]
 mod append;
